Function sys_util::kernel_has_memfd

pub fn kernel_has_memfd() -> bool

Checks if the kernel we are running on has memfd_create. It was introduced in 3.17. Only to be used from tests to prevent running on ancient kernels that won't support the functionality anyways.