Trait sys_util::scoped_signal_handler::SignalHandler

pub unsafe trait SignalHandler {
    fn handle_signal(signal: Signal);
}

The interface used by Scoped Signal handler.

Safety

The implementation of handle_signal needs to be async signal-safe.

NOTE: panics are caught when possible because a panic inside ffi is undefined behavior.

Required methods

fn handle_signal(signal: Signal)

A function that is called to handle the passed signal.

Loading content...