Struct sys_util::signal::KillOnDrop

pub struct KillOnDrop { /* fields omitted */ }

Wraps a Child process, and calls kill_tree for its process group to clean it up when dropped.

Implementations

impl KillOnDrop

pub fn timeout(&self) -> Duration

Get the timeout. See timeout_mut() for more details.

pub fn timeout_mut(&mut self) -> &mut Duration

Change the timeout for how long child processes are waited for before the process group is forcibly killed.