Struct sys_util::ExternalMapping

pub struct ExternalMapping { /* fields omitted */ }

ExternalMapping wraps an external library mapping. This is useful in cases where where the device memory is not compatible with the mmap interface, such as Vulkan VkDeviceMemory in the non-exportable case or when exported as an opaque fd.

Implementations

impl ExternalMapping

pub unsafe fn new(
resource_id: u32,
map: Map,
unmap: Unmap
) -> Result<ExternalMapping>

Creates an ExternalMapping given a library-specific resource id and map/unmap functions.

Safety

The map function must return a valid host memory region. In addition, callers of the function must guarantee that the map and unmap functions are thread-safe, never return a region overlapping already Rust referenced-data, and the backing store of the resource doesn’t disappear before the unmap function is called.