Struct sys_util::PollContext

pub struct PollContext<T> { /* fields omitted */ }

Used to poll multiple objects that have file descriptors.

Example

    let evt1 = EventFd::new()?;
    let evt2 = EventFd::new()?;
    evt2.write(1)?;

    let ctx: PollContext<u32> = PollContext::new()?;
    ctx.add(&evt1, 1)?;
    ctx.add(&evt2, 2)?;

    let pollevents: PollEvents<u32> = ctx.wait()?;
    let tokens: Vec<u32> = pollevents.iter_readable().map(|e| e.token()).collect();
    assert_eq!(&tokens[..], &[2]);

Implementations

impl<T: PollToken> PollContext<T>

pub fn new() -> Result<PollContext<T>>

Creates a new PollContext.

pub fn build_with(fd_tokens: &[(&dyn AsRawFd, T)]) -> Result<PollContext<T>>

Creates a new PollContext and adds the slice of fd and token tuples to the new context.

This is equivalent to calling new followed by add_many. If there is an error, this will return the error instead of the new context.

pub fn add_many(&self, fd_tokens: &[(&dyn AsRawFd, T)]) -> Result<()>

Adds the given slice of fd and token tuples to this context.

This is equivalent to calling add with each fd and token. If there are any errors, this method will stop adding fds and return the first error, leaving this context in a undefined state.

pub fn add(&self, fd: &dyn AsRawFd, token: T) -> Result<()>

Adds the given fd to this context and associates the given token with the fd's readable events.

A fd can only be added once and does not need to be kept open. If the fd is dropped and there were no duplicated file descriptors (i.e. adding the same descriptor with a different FD number) added to this context, events will not be reported by wait anymore.

pub fn add_fd_with_events(
&self,
fd: &dyn AsRawFd,
events: WatchingEvents,
token: T
) -> Result<()>

Adds the given fd to this context, watching for the specified events and associates the given 'token' with those events.

A fd can only be added once and does not need to be kept open. If the fd is dropped and there were no duplicated file descriptors (i.e. adding the same descriptor with a different FD number) added to this context, events will not be reported by wait anymore.

pub fn modify(
&self,
fd: &dyn AsRawFd,
events: WatchingEvents,
token: T
) -> Result<()>

If fd was previously added to this context, the watched events will be replaced with events and the token associated with it will be replaced with the given token.

pub fn delete(&self, fd: &dyn AsRawFd) -> Result<()>

Deletes the given fd from this context.

If an fd's token shows up in the list of hangup events, it should be removed using this method or by closing/dropping (if and only if the fd was never dup()'d/fork()'d) the fd. Failure to do so will cause the wait method to always return immediately, causing ~100% CPU load.

pub fn wait(&self) -> Result<PollEvents<'_, T>>

Waits for any events to occur in FDs that were previously added to this context.

The events are level-triggered, meaning that if any events are unhandled (i.e. not reading for readable events and not closing for hungup events), subsequent calls to wait will return immediately. The consequence of not handling an event perpetually while calling wait is that the callers loop will degenerated to busy loop polling, pinning a CPU to ~100% usage.

Panics

Panics if the returned PollEvents structure is not dropped before subsequent wait calls.

pub fn wait_timeout(&self, timeout: Duration) -> Result<PollEvents<'_, T>>

Like wait except will only block for a maximum of the given timeout.

This may return earlier than timeout with zero events if the duration indicated exceeds system limits.