Struct sys_util::SharedMemory

pub struct SharedMemory { /* fields omitted */ }

A shared memory file descriptor and its size.

Implementations

impl SharedMemory

pub fn named<T: Into<Vec<u8>>>(name: T) -> Result<SharedMemory>

Convenience function for SharedMemory::new that is always named and accepts a wide variety of string-like types.

Note that the given name may not have NUL characters anywhere in it, or this will return an error.

pub fn anon() -> Result<SharedMemory>

Convenience function for SharedMemory::new that has an arbitrary and unspecified name.

pub fn new(name: Option<&CStr>) -> Result<SharedMemory>

Creates a new shared memory file descriptor with zero size.

If a name is given, it will appear in /proc/self/fd/<shm fd> for the purposes of debugging. The name does not need to be unique.

The file descriptor is opened with the close on exec flag and allows memfd sealing.

pub fn from_file(file: File) -> Result<SharedMemory>

Constructs a SharedMemory instance from a File that represents shared memory.

The size of the resulting shared memory will be determined using File::seek. If the given file's size can not be determined this way, this will return an error.

pub fn get_seals(&self) -> Result<MemfdSeals>

Gets the memfd seals that have already been added to this.

This may fail if this instance was not constructed from a memfd.

pub fn add_seals(&mut self, seals: MemfdSeals) -> Result<()>

Adds the given set of memfd seals.

This may fail if this instance was not constructed from a memfd with sealing allowed or if the seal seal (F_SEAL_SEAL) bit was already added.

pub fn size(&self) -> u64

Gets the size in bytes of the shared memory.

The size returned here does not reflect changes by other interfaces or users of the shared memory file descriptor..

pub fn set_size(&mut self, size: u64) -> Result<()>

Sets the size in bytes of the shared memory.

Note that if some process has already mapped this shared memory and the new size is smaller, that process may get signaled with SIGBUS if they access any page past the new size.

pub fn read_name(&self) -> Result<String>

Reads the name from the underlying file as a String.

If the underlying file was not created with SharedMemory::new or with memfd_create, the results are undefined. Because this returns a String, the name's bytes are interpreted as utf-8.