Struct sys_util::TimerFd

pub struct TimerFd(_);

A safe wrapper around a Linux timerfd (man 2 timerfd_create).

Implementations

impl TimerFd

pub fn new() -> Result<TimerFd>

Creates a new timerfd. The timer is initally disarmed and must be armed by calling reset.

pub fn try_clone(&self) -> Result<TimerFd, Error>

Creates a new TimerFd instance that shares the same underlying File as the existing TimerFd instance.

pub fn reset(&self, dur: Duration, interval: Option<Duration>) -> Result<()>

Sets the timer to expire after dur. If interval is not None it represents the period for repeated expirations after the initial expiration. Otherwise the timer will expire just once. Cancels any existing duration and repeating interval.

pub fn wait(&self) -> Result<u64>

Waits until the timer expires. The return value represents the number of times the timer has expired since the last time wait was called. If the timer has not yet expired once this call will block until it does.

pub fn is_armed(&self) -> Result<bool>

Returns true if the timer is currently armed.

pub fn clear(&self) -> Result<()>

Disarms the timer.

pub fn resolution() -> Result<Duration>

Returns the resolution of timers on the host.