Trait sys_util::FromRawDescriptor

pub trait FromRawDescriptor {
    unsafe fn from_raw_descriptor(descriptor: RawDescriptor) -> Self;
}

Required methods

unsafe fn from_raw_descriptor(descriptor: RawDescriptor) -> Self

Safety

Safe only if the caller ensures nothing has access to the descriptor after passing it to from_raw_descriptor